fix: remove record in favor of envoy_log
