+++ kmamiz_data_processor/src/data/mod.rs
@@ -5,7 +5,6 @@ pub mod endpoint_dependency;
 pub mod endpoint_info;
 pub mod historical_data;
 pub mod realtime_data;
-pub mod record;
 pub mod replica_count;
 pub mod request_type;
 pub mod envoy_log;
